requirements.txt
Django==3.1.8
wagtail==2.12.4
gunicorn==20.0.4
dj-database-url==0.5.0
psycopg2>=2.7,<3.0
django-allauth>=0.42.0,<0.43.0
django-debug-toolbar==2.2.1
wagtail-django-recaptcha==1.0
django-crispy-forms==1.11.2
ptvsd == 4.3.2
coverage == 5.2.1
whitenoise == 5.2.0
django-richtextfield == 1.6.0
django-cities-light == 3.8.1
factory-boy == 3.2.0
boto3 == 1.17.56
django-storages == 1.11.1
django-tinymce == 3.3.0
bleach == 3.3.0
python-magic == 0.4.22
django-versatileimagefield == 2.0.0
Pillow == 8.1.1
